+++ src/Service/ResourceService.php
@@ -15,6 +15,7 @@ class ResourceService implements ResourceServiceInterface
         private readonly CacheInterface $resourceCache,
         private readonly SerializerInterface $serializer,
         private readonly MetricsHelper $metricsHelper,
+        private readonly array $excludesResources,
     ) {
     }
 
@@ -35,7 +36,7 @@ class ResourceService implements ResourceServiceInterface
 
         $cachedResources = $this->resourceCache->get("resources-$permission", function (CacheItemInterface $cacheItem) use ($cacheLifetime, $permission) {
             $cacheItem->expiresAfter($cacheLifetime);
-            $info = $this->aakResourceRepository->getAllByPermission($permission);
+            $info = $this->aakResourceRepository->getAllByPermission($permission, $this->excludesResources);
 
             return $this->serializer->serialize($info, 'json', ['groups' => 'minimum']);
         });
@@ -49,7 +50,7 @@ class ResourceService implements ResourceServiceInterface
     {
         $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::INVOKE);
 
-        $info = $this->aakResourceRepository->getOnlyWhitelisted($permission, $whitelistKey);
+        $info = $this->aakResourceRepository->getOnlyWhitelisted($permission, $whitelistKey, $this->excludesResources);
         $serializedWhitelistedResources = $this->serializer->serialize($info, 'json', ['groups' => 'minimum']);
 
         $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::COMPLETE);